+++ CHANGELOG.md
@@ -19,6 +19,8 @@ For more details or to discuss releases, please visit the
   - add nodeCheckboxInput widget
   - round numbers in places
   - color digital values blue when ON
+  - sort nodes by description
+- fix windows build
 
 ## [[0.0.16] - 2021-02-08](https://github.com/simpleiot/simpleiot/releases/tag/v0.0.16)